LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_put5-expected.txt


PASS IDBObjectStore.put() - object store's key path is an object attribute
